refactor code to use es6 modules
