+++ modules/displayPopup.js
@@ -1,0 +1,59 @@
+import commentCount from './commentCount.js';
+import displayComments from './displayComments.js';
+import fetchComments from './fetchComment.js';
+import fetchFishDetails from './fetchFishDetails.js';
+import generatePopup from './generatePopup.js';
+import submitComments from './handleCommentSubmit.js';
+
+const displayPopup = async (fishDetailUrl, fish, name, commentPopup, appId) => {
+  const commentsUrl = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}/comments?item_id=${name}`;
+  const comments = (await fetchComments(commentsUrl)) || [];
+
+  const fishDetails = await fetchFishDetails(
+    fishDetailUrl,
+    fish['Species Name'],
+  );
+
+  // function to get the image link from the api data
+  const fishArray = fishDetails[0]['Image Gallery'];
+  let imageSrc = '';
+  if (fishDetails[0]['Image Gallery']) {
+    imageSrc = fishArray instanceof Array
+      ? fishDetails[0]['Image Gallery'][0]?.src
+      : fishDetails[0]['Image Gallery'].src;
+  }
+
+  // logic to show the popup fo a particular image
+  commentPopup.classList.add('active');
+  generatePopup(commentPopup, fishDetails, imageSrc);
+
+  // function to close the popup when the close icon is clicked
+  const closePopup = document.querySelector('.close-popup');
+  closePopup.addEventListener('click', () => {
+    commentPopup.classList.remove('active');
+  });
+  // Function to submit a comment
+
+  const commentUrl = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}/comments`;
+  const commentsContainer = document.getElementById('comment-list');
+  const usernameInput = document.getElementById('name');
+  const commentInput = document.getElementById('comment');
+  const submitComment = document.getElementById('form');
+  commentsContainer.innerHTML = '';
+
+  // function to display the comments on the dom
+  displayComments(comments, commentsContainer);
+
+  // function to count and display the number of comments for a fish
+  commentCount(commentsContainer);
+
+  // Submit button click event handler
+  submitComment.addEventListener('submit', (e) => {
+    e.preventDefault();
+    const username = usernameInput.value;
+    const comment = commentInput.value;
+    submitComments(name, commentUrl, appId, username, comment);
+  });
+  // end
+};
+export default displayPopup;
